requirements.txt
Flask
Pillow
scikit-learn
torch
transformers
requests
google-generativeai
werkzeug
gunicorn